Rename whole application from MyApp to AuthApp
