+++ android/app/src/main/java/com/authapp/MainActivity.java
@@ -1,4 +1,4 @@
-package com.myapp;
+package com.authapp;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactPackage;
@@ -15,7 +15,7 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "MyApp";
+        return "AuthApp";
     }
 
     /**
