+++ android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'MyApp'
+rootProject.name = 'AuthApp'
 
 include ':app'
