+++ ios/AuthApp/AppDelegate.h
