+++ ios/AuthApp/Images.xcassets/AppIcon.appiconset/Contents.json
