+++ ios/AuthApp/Info.plist
