+++ ios/AuthAppTests/AuthAppTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface MyAppTests : XCTestCase
+@interface AuthAppTests : XCTestCase
 
 @end
 
-@implementation MyAppTests
+@implementation AuthAppTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
