+++ ios/AuthAppTests/Info.plist
